Move logs retrieval code from pcluster_api to commands
