+++ cli/src/pcluster/cli/commands/cluster.py
@@ -11,11 +11,15 @@ import logging
 import os
 import re
 import textwrap
+import time
+from datetime import datetime
 from typing import List
 
 import argparse
 from argparse import ArgumentParser, Namespace
+from tabulate import tabulate
 
+from pcluster import utils
 from pcluster.api.models.cluster_status import ClusterStatusEnum
 from pcluster.cli.commands.common import (
     CliCommand,
@@ -26,6 +30,7 @@ from pcluster.cli.commands.common import (
     print_json,
 )
 from pcluster.constants import PCLUSTER_VERSION_TAG
+from pcluster.models.cluster import Cluster
 from pcluster.validators.common import FailureLevel
 
 LOGGER = logging.getLogger(__name__)
@@ -373,7 +378,7 @@ class ExportClusterLogsCommand(CliCommand):
 
     # CLI
     name = "export-cluster-logs"
-    help = "Export the logs of the cluster to a local tar archive by passing through an Amazon S3 Bucket."
+    help = "Export the logs of the cluster to a local tar.gz archive by passing through an Amazon S3 Bucket."
     description = help
 
     def __init__(self, subparsers):
@@ -426,9 +431,46 @@ class ExportClusterLogsCommand(CliCommand):
         )
 
     def execute(self, args: Namespace, extra_args: List[str]) -> None:  # noqa: D102
-        from pcluster.cli_commands.commands import export_cluster_logs
+        try:
+            output_file_path = args.output or os.path.realpath(
+                f"{args.cluster_name}-logs-{datetime.now().timestamp()}.tar.gz"
+            )
+            self._validate_command_args(output_file_path)
+            self._export_cluster_logs(args, output_file_path)
+        except Exception as e:
+            utils.error(f"Unable to export cluster's logs.\n{e}")
 
-        export_cluster_logs(args)
+    @staticmethod
+    def _validate_command_args(output_file_path: str):
+        # Verify that a file can be written to the given path
+        file_dir = os.path.dirname(output_file_path)
+        if not os.path.isdir(file_dir):
+            try:
+                os.makedirs(file_dir)
+            except Exception as e:
+                utils.error(f"Failed to create parent directory {file_dir} for cluster's logs archive. Reason: {e}")
+        if not os.access(file_dir, os.W_OK):
+            utils.error(f"Cannot write cluster's log archive to {output_file_path}. {file_dir} is not writeable.")
+
+    @staticmethod
+    def _export_cluster_logs(args: Namespace, output_file_path: str):
+        """Export the logs associated to the cluster."""
+        LOGGER.info("Beginning export of logs for the cluster: %s", args.cluster_name)
+        LOGGER.debug("CLI args: %s", str(args))
+        if args.region:
+            os.environ["AWS_DEFAULT_REGION"] = args.region
+
+        cluster = Cluster(args.cluster_name)
+        cluster.export_logs(
+            output=output_file_path,
+            bucket=args.bucket,
+            bucket_prefix=args.bucket_prefix,
+            keep_s3_objects=args.keep_s3_objects,
+            start_time=args.start_time,
+            end_time=args.end_time,
+            filters=" ".join(args.filters) if args.filters else None,
+        )
+        LOGGER.info("Cluster's logs exported correctly to %s.", output_file_path)
 
 
 class ListClusterLogsCommand(CliCommand):
@@ -459,15 +501,47 @@ class ListClusterLogsCommand(CliCommand):
         parser.add_argument("--next-token", help="Token for paginated requests")
 
     def execute(self, args: Namespace, extra_args: List[str]) -> None:  # noqa: D102
-        from pcluster.cli_commands.commands import list_cluster_logs
+        try:
+            self._list_cluster_logs(args)
+        except Exception as e:
+            utils.error(f"Unable to list cluster's logs.\n{e}")
 
-        list_cluster_logs(args)
+    @staticmethod
+    def _list_cluster_logs(args: Namespace):
+        if args.region:
+            os.environ["AWS_DEFAULT_REGION"] = args.region
+
+        cluster = Cluster(args.cluster_name)
+        response = cluster.list_logs(
+            filters=" ".join(args.filters) if args.filters else None,
+            next_token=args.next_token,
+        )
+        if not response.get("logStreams", None):
+            print("No logs found.")
+        else:
+            output_headers = {
+                "logStreamName": "Log Stream Name",
+                "firstEventTimestamp": "First Event",
+                "lastEventTimestamp": "Last Event",
+            }
+            filtered_result = []
+            for item in response.get("logStreams", []):
+                filtered_item = {}
+                for key, output_key in output_headers.items():
+                    value = item.get(key)
+                    if key.endswith("Timestamp"):
+                        value = utils.timestamp_to_isoformat(value)
+                    filtered_item[output_key] = value
+                filtered_result.append(filtered_item)
+            LOGGER.info(tabulate(filtered_result, headers="keys", tablefmt="plain"))
+            if response.get("nextToken", None):
+                LOGGER.info("\nnextToken is: %s", response["nextToken"])
 
 
 class _FiltersArg:
     """Class to implement regex parsing for filters parameter."""
 
-    def __init__(self, accepted_filters):
+    def __init__(self, accepted_filters: list):
         filter_regex = rf"Name=({'|'.join(accepted_filters)}),Values=[\w\-_.,]+"
         self._pattern = re.compile(fr"^({filter_regex})(\s+{filter_regex})*$")
 
@@ -477,7 +551,7 @@ class _FiltersArg:
         return value
 
 
-class GetClusterLogsEventsCommand(CliCommand):
+class GetClusterLogEventsCommand(CliCommand):
     """Implement pcluster get-cluster-log-events command."""
 
     # CLI
@@ -495,7 +569,6 @@ class GetClusterLogsEventsCommand(CliCommand):
             help="Log stream name, as reported by 'pcluster list-cluster-logs' command",
             required=True,
         )
-
         # Filters
         parser.add_argument(
             "--start-time",
@@ -516,7 +589,6 @@ class GetClusterLogsEventsCommand(CliCommand):
         parser.add_argument("--head", help="Gets the first <head> lines of the log stream", type=int)
         parser.add_argument("--tail", help="Gets the last <tail> lines of the log stream", type=int)
         parser.add_argument("--next-token", help="Token for paginated requests")
-
         # Stream utilities
         parser.add_argument(
             "--stream",
@@ -528,6 +600,70 @@ class GetClusterLogsEventsCommand(CliCommand):
         parser.add_argument("--stream-period", help="Sets the streaming period. Default is 5 seconds", type=int)
 
     def execute(self, args: Namespace, extra_args: List[str]) -> None:  # noqa: D102
-        from pcluster.cli_commands.commands import get_cluster_log_events
+        try:
+            self._validate_args(args)
+            self._get_cluster_log_events(args)
+        except Exception as e:
+            utils.error(f"Unable to get cluster's log events.\n{e}")
 
-        get_cluster_log_events(args)
+    @staticmethod
+    def _validate_args(args: Namespace):
+        if args.head and args.tail:
+            utils.error("Parameters validation error: 'tail' and 'head' options cannot be set at the same time")
+
+        if args.stream:
+            if args.next_token:
+                utils.error(
+                    "Parameters validation error: 'stream' and 'next-token' options cannot be set at the same time"
+                )
+            if args.head:
+                utils.error("Parameters validation error: 'stream' and 'head' options cannot be set at the same time")
+        else:
+            if args.stream_period:
+                utils.error("Parameters validation error: 'stream-period' can be used only with 'stream' option")
+
+    def _get_cluster_log_events(self, args: Namespace):
+        """Get log events for a specific log stream of the cluster saved in CloudWatch."""
+        if args.region:
+            os.environ["AWS_DEFAULT_REGION"] = args.region
+        kwargs = {
+            "log_stream_name": args.log_stream_name,
+            "start_time": args.start_time,
+            "end_time": args.end_time,
+            "start_from_head": args.head is not None,
+            "limit": args.head or args.tail or None,
+            "next_token": args.next_token,
+        }
+        cluster = Cluster(args.cluster_name)
+        response = cluster.get_log_events(**kwargs)
+        self._print_log_events(response.get("events", []))
+
+        if args.stream:
+            # stream content
+            next_token = response.get("nextForwardToken", None)
+            while next_token is not None and args.stream:
+                LOGGER.debug("NextToken is %s", next_token)
+                period = args.stream_period or 5
+                LOGGER.debug("Waiting other %s seconds...", period)
+                time.sleep(period)
+
+                kwargs["next_token"] = next_token
+                result = cluster.get_log_events(**kwargs)
+                next_token = result.get("nextForwardToken", None)
+                self._print_log_events(result.get("events", []))
+        else:
+            LOGGER.info("\nnextBackwardToken is: %s", response["nextBackwardToken"])
+            LOGGER.info("nextForwardToken is: %s", response["nextForwardToken"])
+
+    @staticmethod
+    def _print_log_events(events: list):
+        """
+        Print given events.
+
+        :param events: list of boto3 events
+        """
+        if not events:
+            print("No events found.")
+        else:
+            for event in events:
+                print("{0}: {1}".format(utils.timestamp_to_isoformat(event["timestamp"]), event["message"]))
