+++ cli/src/pcluster/cli_commands/commands.py
@@ -16,12 +16,9 @@
 import logging
 import os
 import sys
-import time
 from builtins import str
-from datetime import datetime
 
 from botocore.exceptions import ClientError
-from dateutil import tz
 from tabulate import tabulate
 
 import pcluster.utils as utils
@@ -436,155 +433,6 @@ def update(args):
     except KeyboardInterrupt:
         LOGGER.info("\nExiting...")
         sys.exit(0)
-
-
-def export_cluster_logs(args):
-    """Export the logs associated to the cluster."""
-    LOGGER.info("Beginning export of logs for the cluster: %s", args.cluster_name)
-    LOGGER.debug("CLI args: %s", str(args))
-
-    try:
-        # Verify that a file can be written to the given path
-        output = args.output or os.path.realpath(f"{args.cluster_name}-logs-{datetime.now().timestamp()}.tar.gz")
-        file_dir = os.path.dirname(output)
-        if not os.path.isdir(file_dir):
-            try:
-                os.makedirs(file_dir)
-            except Exception as exception:
-                utils.error(
-                    f"Failed to create parent directory {file_dir} for cluster's logs archive. Reason: {exception}"
-                )
-        if not os.access(file_dir, os.W_OK):
-            utils.error(f"Cannot write cluster's log archive to {output}. {file_dir} is not writeable.")
-
-        # export logs cluster raises an exception if stack does not exist or for other internal errors
-        result = PclusterApi().export_cluster_logs(
-            cluster_name=args.cluster_name,
-            region=utils.get_region(),
-            output=output,
-            bucket=args.bucket,
-            bucket_prefix=args.bucket_prefix,
-            keep_s3_objects=args.keep_s3_objects,
-            start_time=args.start_time,
-            end_time=args.end_time,
-            filters=" ".join(args.filters) if args.filters else None,
-        )
-        if isinstance(result, ApiFailure):
-            utils.error(f"Unable to export cluster's logs.\n{result.message}")
-        else:
-            LOGGER.info("Cluster's logs exported correctly to %s.", output)
-    except KeyboardInterrupt:
-        LOGGER.info("Exiting...")
-        sys.exit(0)
-
-
-def list_cluster_logs(args):
-    """List available cluster logs."""
-    try:
-        result = PclusterApi().list_cluster_logs(
-            cluster_name=args.cluster_name,
-            region=get_region(),
-            filters=" ".join(args.filters) if args.filters else None,
-            next_token=args.next_token,
-        )
-        if isinstance(result, ApiFailure):
-            utils.error(f"Unable to retrieve the list of cluster's logs.\n{result.message}")
-        elif not result.get("logStreams", None):
-            print("No logs found.")
-        else:
-            output_headers = {
-                "logStreamName": "Log Stream Name",
-                "firstEventTimestamp": "First Event",
-                "lastEventTimestamp": "Last Event",
-            }
-            filtered_result = []
-            for item in result.get("logStreams", []):
-                filtered_item = {}
-                for key, output_key in output_headers.items():
-                    value = item.get(key)
-                    if key.endswith("Timestamp"):
-                        value = utils.timestamp_to_isoformat(value)
-                    filtered_item[output_key] = value
-                filtered_result.append(filtered_item)
-            LOGGER.info(tabulate(filtered_result, headers="keys", tablefmt="plain"))
-            if result.get("nextToken", None):
-                LOGGER.info(f"\nnextToken is: {result['nextToken']}")
-    except KeyboardInterrupt:
-        LOGGER.info("Exiting...")
-        sys.exit(0)
-
-
-def get_cluster_log_events(args):
-    """Get log events for a specific log stream of the cluster saved in CloudWatch."""
-
-    if args.head and args.tail:
-        utils.error("Parameters validation error: 'tail' and 'head' options cannot be set at the same time")
-
-    if args.stream:
-        if args.next_token:
-            utils.error("Parameters validation error: 'stream' and 'next-token' options cannot be set at the same time")
-        if args.head:
-            utils.error("Parameters validation error: 'stream' and 'head' options cannot be set at the same time")
-    else:
-        if args.stream_period:
-            utils.error("Parameters validation error: 'stream-period' can be used only with 'stream' option")
-
-    limit = None
-    start_from_head = False
-    if args.head:
-        limit = args.head
-        start_from_head = True
-    elif args.tail:
-        limit = args.tail
-
-    try:
-        kwargs = {
-            "cluster_name": args.cluster_name,
-            "region": get_region(),
-            "log_stream_name": args.log_stream_name,
-            "start_time": args.start_time,
-            "end_time": args.end_time,
-            "start_from_head": start_from_head,
-            "limit": limit,
-            "next_token": args.next_token,
-        }
-        result = PclusterApi().get_cluster_log_events(**kwargs)
-        if isinstance(result, ApiFailure):
-            utils.error(f"Unable to retrieve the log events.\n{result.message}")
-        else:
-            _print_log_events(result.get("events", []))
-            if args.stream:
-                # stream content
-                next_token = result.get("nextForwardToken", None)
-                while next_token is not None and args.stream:
-                    LOGGER.debug(f"NextToken is {next_token}")
-                    period = args.stream_period or 5
-                    LOGGER.debug(f"Waiting other {period} seconds...")
-                    time.sleep(period)
-
-                    kwargs["next_token"] = next_token
-                    result = PclusterApi().get_cluster_log_events(**kwargs)
-                    next_token = result.get("nextForwardToken", None)
-                    _print_log_events(result.get("events", []))
-            else:
-                LOGGER.info(f"\nnextBackwardToken is: {result['nextBackwardToken']}")
-                LOGGER.info(f"nextForwardToken is: {result['nextForwardToken']}")
-    except KeyboardInterrupt:
-        LOGGER.info("Exiting...")
-        sys.exit(0)
-
-
-def _print_log_events(events: list):
-    """
-    Print given events.
-
-    :param events: list of boto3 events
-    """
-    if not events:
-        print("No events found.")
-    else:
-        for event in events:
-            print("{0}: {1}".format(utils.timestamp_to_isoformat(event["timestamp"]), event["message"]))
 
 
 def delete_image(args):
